Fix attach urls
