+++ superannotate/__init__.py
@@ -50,7 +50,8 @@ from .db.images import (
     download_image_preannotations, get_image_annotations, get_image_bytes,
     get_image_metadata, get_image_preannotations, search_images,
     search_images_all_folders, set_image_annotation_status,
-    set_images_annotation_statuses, upload_image_annotations, get_project_root_folder_id
+    set_images_annotation_statuses, upload_image_annotations,
+    get_project_root_folder_id
 )
 from .db.project_api import (
     create_folder, delete_folders, get_folder_metadata,
