+++ superannotate/db/search_projects.py
@@ -27,7 +27,7 @@ def search_projects(
     params = {
         'team_id': str(_api.team_id),
         'offset': 0,
-        'limit' : limit,
+        'limit': limit,
         'completeImagesCount': include_complete_image_count
     }
     if name is not None:
